[General] Refactor backend to prepare for backend integration

+++ application/LocalProfile.py
@@ -2,7 +2,7 @@ import copy
 import json
 import uuid
 
-from application.paths import *
+from .paths import *
 
 _PROFILE_VERSION = 1  # in case the schema changes in the future
 
@@ -14,11 +14,10 @@ _EMPTY_SESSION = {
 _EMPTY_USER_PROFILE = {
     "version": _PROFILE_VERSION,
     "sessions": {},
-    "last_session": ""
 }
 
 
-class Profile:
+class LocalProfile:
     def __init__(self):
         self._profile = copy.deepcopy(_EMPTY_USER_PROFILE)
         try:
@@ -26,60 +25,72 @@ class Profile:
                 json_data = json.load(infile)
 
                 if "version" not in json_data or json_data["version"] != _PROFILE_VERSION:
-                    raise ValueError("Profile: Version info not found or mismatch in the profile.")
+                    raise ValueError("LocalProfile: Version info not found or mismatch in the profile.")
 
-                # FIXME: make sure the profile file is not modified in an attempt to crash the script
-                self["sessions"] = json_data["sessions"]
-                self["last_session"] = json_data["last_session"]
+                # TODO: make sure the sessions file is not modified in an attempt to crash the script
+                self._profile["sessions"] = json_data["sessions"]
+                self._profile["last_session"] = json_data["last_session"]
 
         except Exception as e:
             self._profile = copy.deepcopy(_EMPTY_USER_PROFILE)
-            print("Profile: load_profile:", e)
+            print("LocalProfile: load_profile:", e)
             # raise e
             print("Unable to load the user profile. Using the default profile instead.")
 
-    def __setitem__(self, key, value):
-        self._profile[key] = value
-
-    def __getitem__(self, key):
-        return self._profile[key]
+    def query(self):
+        return self._profile
 
     # TODO: support renaming a session
     def add_session(self, host, username):
         session = copy.deepcopy(_EMPTY_SESSION)
 
         session_id = uuid.uuid4().hex
-        self["sessions"][session_id] = session
+        self._profile["sessions"][session_id] = session
         session["host"] = host
         session["username"] = username
 
-        # TODO: check whether we should have a dedicated function to update the last session
-        self["last_session"] = session_id
         self.save_profile()
 
         return session_id
 
     def delete_session(self, session_id):
+        if session_id not in self._profile['sessions']:
+            return False, f'failed: session {session_id} does not exist'
+
         try:
             os.remove(os.path.join(PRIVATE_KEY_PATH, session_id))
         except FileNotFoundError:
             print('Not valid SSH key found for deletion')
 
-        self['sessions'].pop(session_id)
+        self._profile['sessions'].pop(session_id)
         self.save_profile()
 
+        return True, ''
+
     def change_host(self, session_id, new_host):
-        self["sessions"][session_id]['host'] = new_host
+        if session_id not in self._profile['sessions']:
+            return False, f'failed: session {session_id} does not exist'
+
+        self._profile["sessions"][session_id]['host'] = new_host
         self.save_profile()
+
+        return True, ''
 
     def save_profile(self):
         try:
             with open(USER_PROFILE_PATH, 'w') as outfile:
                 json_data = json.dumps(self._profile, indent=4)
                 outfile.write(json_data)
-        except Exception:
+        except Exception as e:
             # need to handle any write permission issues, once observed
-            raise Exception
+            raise e
 
-    def query(self):
-        return self._profile
+    def get_session_info(self, session_id):
+        if session_id not in self._profile['sessions']:
+            return None, None, None
+
+        host = self._profile['sessions'][session_id]['host']
+        username = self._profile['sessions'][session_id]['username']
+        this_private_key_path = os.path.join(PRIVATE_KEY_PATH, session_id)
+
+        return host, username, this_private_key_path
